refactor: consolidate meta_analysis micro-files into 3 modules
Merge 18 single-function files into field_mapping.py, reporting.py,
and sarif_analysis.py. Keep locations_match.py separate (181 lines).
Update all imports in cli/inspect/ and tests.
