+++ automated_security_helper/cli/inspect/sarif_fields.py
@@ -9,10 +9,10 @@ import os
 from pathlib import Path
 from typing import Annotated, Dict, Any
 
-from automated_security_helper.utils.meta_analysis.generate_field_mapping_html_report import (
+from automated_security_helper.utils.meta_analysis.reporting import (
     generate_html_report,
 )
-from automated_security_helper.utils.meta_analysis.should_include_field import (
+from automated_security_helper.utils.meta_analysis.field_mapping import (
     should_include_field,
 )
 
@@ -136,22 +136,6 @@ def analyze_sarif_fields(
             help="Directory to write output files",
         ),
     ] = None,
-    # config: Annotated[
-    #     Optional[str],
-    #     typer.Option(
-    #         "--config",
-    #         "-c",
-    #         help="The path to the configuration file",
-    #         envvar="ASH_CONFIG",
-    #     ),
-    # ] = None,
-    # config_overrides: Annotated[
-    #     List[str],
-    #     typer.Option(
-    #         "--config-overrides",
-    #         help="Configuration overrides specified as key-value pairs (e.g., 'reporters.cloudwatch-logs.options.aws_region=us-west-2')",
-    #     ),
-    # ] = [],
 ):
     """
     Analyze SARIF fields across different scanners to understand their schema.
