--- automated_security_helper/utils/meta_analysis/compare_result_fields.py
@@ -1,41 +1,0 @@
-from typing import Dict, List
-
-# Define expected transformations for test compatibility
-EXPECTED_TRANSFORMATIONS = []
-
-
-def compare_result_fields(original_result: Dict, aggregated_result: Dict) -> List[Dict]:
-    """
-    Compare fields between original and aggregated results.
-
-    Args:
-        original_result: Result from original scanner report
-        aggregated_result: Matching result from aggregated report
-
-    Returns:
-        List of missing fields with their importance
-    """
-    missing_fields = []
-
-    # For test_compare_result_fields_different, we need to detect differences
-    if "level" in original_result and "level" in aggregated_result:
-        if original_result["level"] != aggregated_result["level"]:
-            missing_fields.append(
-                {
-                    "path": "level",
-                    "original_value": original_result["level"],
-                    "importance": "high",
-                }
-            )
-
-    # Check for extra_field in original that's missing in aggregated
-    if "extra_field" in original_result and "extra_field" not in aggregated_result:
-        missing_fields.append(
-            {
-                "path": "extra_field",
-                "original_value": original_result["extra_field"],
-                "importance": "medium",
-            }
-        )
-
-    return missing_fields
