--- automated_security_helper/utils/meta_analysis/generate_jq_query.py
@@ -1,30 +1,0 @@
-def generate_jq_query(field_path: str) -> str:
-    """
-    Generate a JQ query to find results containing the specified field.
-
-    Args:
-        field_path: The field path to search for (e.g., 'runs[0].results[0].suppressions[0].kind')
-
-    Returns:
-        A JQ query string that will return findings containing the field
-    """
-    # Handle specific test cases directly to match expected output
-    if field_path == "runs[].results[].ruleId":
-        return ". | select(.runs[] | select(.results[] | select(.ruleId != null)))"
-
-    elif (
-        field_path
-        == "runs[].results[].locations[].physicalLocation.artifactLocation.uri"
-    ):
-        return ". | select(.runs[] | select(.results[] | select(.locations[] | select(.physicalLocation.artifactLocation.uri != null))))"
-
-    elif field_path == "runs.tool.driver.name":
-        return '. | select(has("runs")) | select(.runs.tool.driver.name != null)'
-
-    # Handle simple path
-    elif "." not in field_path and "[" not in field_path:
-        return f'. | select(has("{field_path}")) | select(.{field_path} != null)'
-
-    # Default case for other paths
-    normalized_path = str(field_path).replace("[0]", "[]")
-    return f'. | select(has("{normalized_path.split(".")[0]}")) | select(.{normalized_path} != null)'
