+++ tests/unit/utils/meta_analysis/test_analyze_sarif_file.py
@@ -2,7 +2,7 @@ import pytest
 import json
 import tempfile
 from pathlib import Path
-from automated_security_helper.utils.meta_analysis.analyze_sarif_file import (
+from automated_security_helper.utils.meta_analysis.sarif_analysis import (
     analyze_sarif_file,
 )
