+++ tests/unit/utils/meta_analysis/test_extract_field_paths.py
@@ -1,4 +1,4 @@
-from automated_security_helper.utils.meta_analysis.extract_field_paths import (
+from automated_security_helper.utils.meta_analysis.field_mapping import (
     extract_field_paths,
 )
