+++ tests/unit/utils/meta_analysis/test_get_message_text.py
@@ -1,4 +1,4 @@
-from automated_security_helper.utils.meta_analysis.get_message_text import (
+from automated_security_helper.utils.meta_analysis.sarif_analysis import (
     get_message_text,
 )
