+++ tests/unit/utils/meta_analysis/test_get_value_from_path_coverage.py
@@ -1,6 +1,6 @@
 """Test coverage for get_value_from_path function."""
 
-from automated_security_helper.utils.meta_analysis.get_value_from_path import (
+from automated_security_helper.utils.meta_analysis.field_mapping import (
     get_value_from_path,
 )
